+++ Source/iti-react/Inputs/RadioInput/RadioInput.tsx
@@ -102,6 +102,6 @@ function required() {
     })
 }
 
-export const RadioInputValidators = {
+export const RadioValidators = {
     required
 }